add favicon, change page title, add article keys, change button color
This commit adds a favicon to the site. Changes the page title to
'Newsweek Clone'. Unique keys were added to the articles by getting
the index of the item from the section article array. The next
button on some browsers by default turned blue after a click so I
set the next button color to be black in css.